Added Test for Arabic to Roman Conversion
I added tests for the aravicToRoman() function. I made tests for the minimum and maximum possible numbers that can be converted into roman numerals. Made sure the function can function when the given integer is not inside of the tree map. Also made sure it can handle integers with one, two, three, or four digits.